use signal for keyboard store + remove
